make private unued public function
