+++ index.html
@@ -5,17 +5,9 @@
 
     <script src="dist/yawl.js"></script>
     <script>
-      yawl.configure({apiKey: 'cda712a73aff22114b6f62871697ea15'})
+      yawl.configure('cda712a73aff22114b6f62871697ea15')
 
-      //yawl.reset()
-      // ahoy.trackAll();
-      // ahoy.trackView({hello: "world"});
-      //yawl.trackClicks("a");
       yawl.track("click");
-
-      // fetch("/fetch", {
-      //   headers: {"Ahoy-Visit": ahoy.getVisitId(), "Ahoy-Visitor": ahoy.getVisitorId()}
-      // })
     </script>
   </head>
   <body>
